parallelize git-conut, optimize for speed
And paging in with multiple threads is way faster, now reaching speeds
of 280MB/s as compared to 80MB on a single core.

Parsing the object is slow, but now multithreaded at least.

It looks like a simple par-iter already does the job, so it's really
worth thinking strongly about whether or not to use async for
everything.

'Blocking' could easily used to unblock the little IO that happens
directly, but there is no way to unblock the page ins - there
one really has to use real threads which usually works well to speed
things up as it's CPU + IO bound, in that order.