First step towards streaming of ZLIB deflated content
Based on deflate2, which is somewhat complex since it supports plenty of
probably rather different backends.

It also comes with additional dependencies, that I am happy to just
avoid for now - how hard can it be to implement streaming decompression
and compression using a BufReader based on code in flate2 - famous
last words ;)?