[url] Now with support for non-utf8 byte strings
Git urls are a hybrid of paths and actual URLs, which have entirely
different encodings.

We manage it, but restrict paths to byte formats, and won't be able
to deal with UTF-16 or the likes.

So if some input like this is about to be encountered, it should really
be put as UTF-8 or something similar.