[clone] unbelievable, but it now seems to work as expected
And I have no idea what the difference is accept that the read patterns
changed.

Let's hope this issue doesn't return or that we soon have an altenrative
HTTP(S) layer implementation that is…deterministic.

It's interesting to see that a supposedly simple interaction with a
Rust-wrapped C-library already shatters my feeling of safety and
adds a lot of time due to weirdness that really wasn't ever
a problem before.