[git-config] Add coercion into owned variants
An ergonomic issue is that we currently require a reference to some
bytes, rather than allow users to have the Parser have owned instances
of all events. This allow users to have a Parser that owns all its
events and can be passed around without the original bytes in memory, if
they would prefer having ergonomics over performance.