Revert "Add additional variant for Sha256 in ObjectId (#63)"
This reverts commit 3dd7c4350e140b72c21598f95a4557e6115d3124.

It's not necessary to have this now as it affects a few performance
relevant structures (memory consumption when indexing).

It's better to do this properly once it's clear how that should be
implemented at all - then we can offer options to not compile
Sha256 support in or solve the memory issue differently, for instance,
by supporting zero-copy for most entries as well or be smart about
how we allocate (like using Arenas).