Introduce pack_id for use in pack cache, preventing (most collisions)
Right now only crc32 is used based on the path, but one could imagine
to use 4 bytes of a SHA1 instead.

Collisions would mean potential reuse of a base object which isn't
actually the one we expect, in the best case this panics, in the worst
it produces an incorrect object which might not be recognized outside
of fsck.

Ideally, the pack-id never clashes, and there are certainly other ways
to make that very very unlikely like a global counter.