Support for caches within linked::Db
Since the only cache supported right now is caching packed objects,
as opposed to loose objects which one day probably needs to be added
as well, namespacing is already supported and packs will simply be
'flattened' from the perspective of the cache. That is, 10 repos
with 2 packs each will yield entries from 10 different packs within
a single cache.