[git-ref] clear it out and move existing functionality to git-object
This removes an uncanny dependency.
Maybe now we have it the other way around.

Possibly there should be a new crate to contain shared validation
code? Maybe it should just be duplicated? No, that could mean code
drift.

It makes sense for the functionality to be included in the crate
that needs it, so lets see what kind of input validation needs to
happen for refs to work and if it truly is the same.

As it's pretty stable there shouldn't be too much worry about somewhat
similar code to exist in various crates.