[git-protocol] fetch in sync and async…
…which also caused the Send requirement to be removed. It was coming in
with the `async_trait` proc macro.

This also means that we learned that Send is possible for everyone
involved and if we want that.

Now only local processing is supported which should be fine as well.

The reason for removing the Send bound? Thanks to maybe-async it would
have forced Send bounds on many blocking implementations as well
as we only have one fetch impl.

In practice it should be fine to deal with that limitation as
processing happens in async-land on one threaded executor, and when
needed the transport wrapper can be moved to a blocking executor
either directly. It's just that our futures are not Send, which is
fine as most threaded executors need 'static as well.