[transport] unsupported protocol versions now abort the fetch operation
We don't close the connection because the borrow checker won't let us,
but that's probably already and happens in other places as well
in case there is an error. Transports probably close on drop, right?

What about async drop?

Maybe we are leaking all the time, unless people use `&mut transport`.