[ref #139] refactor
Reinvent the wheel less - after all we will need access to `git-object`
or delegate all chain following to the caller.

An alternative would be a feature-gated implementation of the resolution
algorithm behind behind function signature that doesn't know git-object,
but it's certainly not worth it.

Let's also think of revparsing, which definitely needs some git-object
access too.