[smart-release #162] FAIL: promising at first, but not really working…
…as multiple fields of the cache can't be borrowed at the same time.
Makes a little sense, too, as we need a shared borrow for packed-refs
and a mutable one for the pack cache. This is non-overlapping but
it seems the RefCell implementation can't handle that.

Right now it compiles but the test fails as the RefMut is held too long.
If that is changed it turns out the value we get is only temporary,
so can't get read-only access to one field while getting write access
to another.