[config #185] add test for handling windows formatted files…
…which occour implicitly when checking out the repository fixtures on
windows, which are then windows formatted. From there the unholy
carriage return \r enters the compiled binary using include_bytes!
as a template, which then is used to initialize a repository.

In case of bare repositories, the implementation has to check if
it really is bare (i.e. bare = true) or just looks bare (because
there is no index, but that's the case in a new repository as well).

This check would use the config parser, which wouldn't consider carriage
returns and fail to parse the config file, making the opening of
repositories impossible on windows.