[repository #190] note a known limitation about finding references in namespaces…
…which doesn't work unless the fully qualified namespaced name is
provided. This is correct behaviour.

To fix this, one would have to search in a known namespace, which
can be done with some refactoring. This would allow refs to be found
by partial name.

However, doing this is only useful for client applications who want
to namespace a repository and allow the user to search by partial name.