fix: '(null)' symref targets are turned into direct refs instead… (#205)
…because that's what 'git' would do if it would care enough.
On the client side this means one has to deal with detached heads
and it's obvious that these are detached. It's not clear anymore
why this is the case, and probably it's good to hide it as the
current git behaviour is accidental and may change in future, and
do so in a way we predict.

There is the possibility that git would abort the entire
fetch/upload-pack operation, but that's already handled correctly
by our implementation as well as we understand ERR messages in packet
lines.