feat: add easy::ext::ObjectAccessExt::tag(…) to create tag objects  (#198)
It's a quick sketch on how tag object creation could work.

Note the duplication the method name using traits, which seems like a good solution
to the problem of differentiating tag objects and tag references while
keeping the method name short.

Most will only ever need one, right?

Even in my example that's not the case, so maybe we have to rename it.