refactor!: Remove --no-multi-crate-release support entirely (#222)
As the default is to do multi-crate releases and now having to deal
with single-create releases reduces maintenance burden.

The solution to this problem is to not specify versions in
dev-dependencies to workspace crates.

We also don't check for this anymore, which might be re-added
at some point if there is demand.