fix: dependency resolution (#222)
Previously the ordering of crate for release might not have been
correct due to this issue that is now fixed.

We need depth-first traversals and previously it would extend skipped
dependencies, effectively putting them into their own ordering.

Previously it would restore that ordering, but not anymore, causing
this bug that was entirely unnecessary.