fix: panic due to unexpected internal state (#222)
When there was no change in the src/ directory of the top-level crate,
the dependency resolution would not be able to auto-bump the version
as no change occurred, but another part would usually detect a change
as it wasn't confined to the top-level src/ directory.

This could lead to a panic as an invariant wasn't upheld.

This was fixed by letting both parts agree to use the src/ directory
to determine changes of the top-level directory, and by making panics
impossible while improving the messaging around this state should it
still occur. The latter is rough, probably rare, but usable.