Don't unnecessarily reload the packed-refs buffer under contention (#263)
Multiple writers can come to the same conclusion and try to reload
a packed buffer. Now they will reassuure themselves this is still needed
after getting exclusive write access.

Notably, git doesn't do that as it uses snapshots, and libgit2 also
doesn't seem to do that. Technically this is a bit slower in the common
case due to the extra state call.