feat: Add InOrderIter to 'parallel' module (#293)
This iterator makes possible identifies results using a sequence id
and returns only consecutive items.

Use it to collect unordered results produced by threads.
It's advantage to collecting yourself and sorting is the potential
for a smaller memory footprint of in-flight results, one doesn't
have to collect them all for ordering, necessarily.