fix: Don't pass judgement on usefulness of certain kinds of git-conventional messages (#318)
Previously we would intentionally avoid writing out information about
refactors or chores as they are not deemed useful in a changelog.

However, this can be confusing for anyone but the original author.

We now write them as seen.

Future iterations on this may consider adding more options
to configure which sections should go into the changelog.