fix: Assure interrupt based tempfile cleanup can't deadlock (#328)
We do this by using the `try_entry()` API provided by the most recent
dashmap, at the cost of potentially trying to access a lot of indices
that don't exist in the map anymore. The cost of this are expected
to be low though especially since interrupts are the uncommmon case.

A side-effect of this is that tempfiles that are currently being
removed for writing to them, for example, won't be cleaned up.

It will be up to the code running after the interrupt, if the case
at all, to deal with the tempfile, which is what it does anyway.