turn PathError into path::interpolate::Error; refactor (#331)
Previously PathError was meant to be used primarily in `Path::try_from()`
so a case could be made to give it that name. Now the error is exclusive
to `Path`, so that should be reflected in the module location as well.

This helps with organizing the code as well, putting everything related
to Path into its own module, which helps refactoring later.