finally an understanding on collision checking (#301)
Let's not try to imitate git here but instead solve the problem our
way while learning from bugs git ran into in the process.

This will be particularly interesting when dealing with parallel
checkouts, something we need to implement for sure.

On the bright side, there seems to be opportunity for a performance
boost due to less iops during clone.