basic version of index checkout via command-line (#301)
For now just with empty files, but that shows that with a single thread
it's faster than git at 27k files per second created compared to 14.5k
of a single thread of git.

We do the least amount of work necessary even without an lstat cache,
but that probably changes once we have to access the ODB to obtain
actual data.

Note that git might also do an additional exists check per file
to see if it changed, something we don't do as we may assume
exclusive access to the directory and just go with that for now.

Long story short: it looks like there is a lot of potential for
performance improvements and I think there is a lot of room for
being faster especially in multi-threaded mode.