refactor include path recursion logic (#331)
It was born in the `from_paths()` method which eventually showed
that it wasn't possible to easily resolve paths in an existing
config file, as is the case in the `from_env()` constructor.

There still is a TODO asking how to resolve relative include paths
if there was no path on disk.