adjust wildmatch corpus expectations as it won't match our preprocessor (#301)
It appears that we do truncate trailing backslashes differently
at least in our test setup, so the actual pattern used for matching
is different and produces different (but correct results.

Hence we update the expectation accordingly, instead of adjusting
the pattern parsing, as it seems unnecessary to match such patterns
or to adjust the parser to support them.