Take a couple more steps to appease the CI gods.
1.) Install gnu-sed on the macOS instances

2.) Remember that which(1) sets a non-zero exit code if it doesn't find
each item you've specified.  In this case we're searching for 'gsed' and
'sed' in the hopes of finding at least one.

The ultimate goal is to prefer gsed if it exists as systems with a BSD
userland tend to install GNU tools with a 'g' prefix such that GNU sed
is accessible as 'gsed'.

GNU sed lets us easily (in a somewhat portable way) edit the second line
of a sequence so that we can change one 'pick' to 'edit'.  All while
pretending to be an interactive text editor to make git happy.