feat: support for `try_value()`, `boolean()` and `string()` access`. (#301)
Support for a convenient way of knowing if a value does or doesn't exist
via `try_value()`, which can only fail if the conversion fails.

Lastly, `string()` is a special case which doesn't fail as there is
no conversion, and `boolean()` allows to obtain a plain boolean value
if it was a valid boolean representation.