fix: Avoid running into the `default-members` trap with 'cargo publish'. (#429)
Default-members in a cargo workspace can override what's actually
published, so we have to be explicit about what to publish.

This is only the case when there is more than one members in the
workspace, even though it would probably work as well if the package
would be specified with a single-crate workspace.