improved slash/backslash handling on windows (#331)
It's very icky, and needs careful comparison to what git actually does
so we can mimick it. When backslashes get into play, things get a bit
'special' as git will sometimes transform the pattern to forward
slashes, but not always.

Every one of our invocations of `to_unix_paths_on_windows()` is
motivated by a test.