reuse the initialized environment for a little speed (#331)
Even though this is technically a little less clean, we are read-only
and really only need a bare repo whose three files we fully control
to setup the test-case.

It's still a bit strange the the `Command` invocations of `git` are
very slow, even though git finishes in 10ms when tried by hand.