fix: support keepalive packets. (#450)
Keepalive packets are side-band only empty datalines that should
just be ignored. This is now happening, allowing longer git
operations to work as they will send keepalive packets every 5 seconds,
and previously we would choke on it.

Note that empty datalines are never send otherwise, making it a
previously unused marker that can safely be skipped.