feat: Add various methods to iterate sections along with their id, and mutate them.
As section names are not unique, it was previously not possible to
iterate sections and then mutate them as one wouldn't be able to refer
to the exact section that was just traversed, after all, there can be
many sections named `remote "origin"`.

With the new methods it's possible to uniquely refer to each section
for mutation and removal.