feat: add `RefSpecRef::expand_prefix()` method to avoid missing prefixes. (#450)
The current implementation might cause refspecs to end up ignored as
they don't have a prefix, and in protocol V2 it would then fail
to add a ref-spec filter which causes them to be missed.

With `expand_prefix()`, we assure that there are all possible prefixes
that can contain partial names, similar to what git does.