allow to re-read the logallrefupdates config after overrides (#450)
It's still not enough though as it won't be able to write ref-logs
when receiving packs, so that kind of override isn't possible
and I wonder if it's worth the effort.

Also, it's not yet possible to enforce reflogs when writing received
refs, only the HEAD, and even that isn't really working yet.