feat: `gix remote ref-map --show-unmapped-remote-refs`. (#450)
That way it's more obvious to see what was filtered out by ref-specs.

It's also great to validate that server-side filtering via ref-prefix
will not send refs that are referred to by symbolic refs that are
not filtered out. That should be fine as it's all about objects,
it's just something to deal with as we may have to deal with symbolic
refs that aren't in the set of refs the server sent to us.