Avoid invalid invocations of `git` by removing `GIT_CONFIG_COUNT`
Thanks to this environment variable invocations of `git upload-pack`
and the likes can fail to startup as they encounter an invalid,
intermediate state as created by unit tests that run in parallel.

As long as there is no legitimate case for it, this should be alright.

Lastly, the goal will be to be self-serving and avoid starting any
git command as we can probably do it faster or with more control
for the specific usecase.