feat!: add `remote::Connection::with_transport_config()`, change the way `*::transport_mut()` is used.
Previously `transport_mut()` was supposed to be used for calling
`configure()`, but that doesn't work anymore as `configure()` can
only effectivey be called once the initialization of the Connection
is complete, as it may depend on the Remote name AND the credential
provider for proxy auth credential acquisition.

Thus we allow callers to set the transport options they need in advance
for it to be used when needed.