change!: move `fetch::refs(…, protocol, …)` to the crate root as `ls_refs(…)`.
This move also removes the `protocol` parameter as ls-refs is only available
in V2 and the caller is forced to know that after running a
`handshake(…)` anyway which yields the actual protocol used by the
server.

Note that the client can suggest the version to use when connecting,
which is specific to the transport at hand.