fix: don't overwrite unexpanded `git-lfs` pointer files. (#607)
It's possible for those with incomplete `git-lfs` installations
(and many more situations) to end up in a spot where pointer files
aren't expanded. If we overwrite the with archives, files look
changed which can be confusing and lead to even bigger messes
to happen.

Now we don't overwrite those files anyomre.