fix!: subsections are identified as `&BStr` in entire API.
Technically they can be any value (except for newlines and unescaped double quotes),
and these values might be paths and everything that comes with it, like
illformed UTF8. In order to be able to represent everything that
git can represent, we don't enforce UTF8 anymore for subsection names.

Note that section names and key names are required to be valid UTF8
(and even alphanumeric ascii), which makes illformed UTF8 very unlikely
there.