feat!: represent object cache configuration like `GITOXIDE_PACK_CACHE_MEMORY` in git-configuration.
That way there is a unified system for how to set values, which may be overridable by configuration
variables or not.

With this changes, the explicit application of environment variables for setting the cache
isn't required anymore as everything happens using git-configuration, and automatically,
while providing full control like before.