fix!: `Capabiltiies::from_lines()` takes a single buffer.
That way it doesn't have to convert to `String` as intermediary
which may fail as illformed UTF8 might be present.

Performance wise, reading all lines ahead of time, it is the same
as it was before as it would collect all lines beforehand anyway.

We are also seeing only a few lines in V2, and in V1 it was
fully streaming already.