change!: environment variable permissions are per topic.
Now `Permissions` for environment variables are so that they
are by topic instead of by prefix, by default. That way
it's easier to allow or deny particular sets of related
environment variables.

The catch-all permissions by prefix are still present for all
other variables that aren't contained in one particular topic.