change!: default features are set to `max-performance-safe` to assure compatibility.
Previously the `max-performance` setting might have caused issues during compilation
or issues at runtime if libraries like `git2` are used in the same binary, and the
new default feature settings maximizes compatbility so this won't happen.

For best performance, however, one will have to activate the `max-performance`
feature on top of that.