fix!: allow to traverse the entire commit graph of shallow repos
Previously, when traversing commits, we would assume to be in a
shallow repository if a commit's parent could not be found in the
repository.

Now we validate that assumption by reading the 'shallow' file to
check if the last seen commit is on the commit boundary.

This removes `is_shallow` and `error_on_missing_commit()` on the
`revision::walk::Platform` as shallow commits are now known and handled
without any guesswork.