Correct more typos with `typos` tool.
The initial configuration still has a few problems:

after basic configuration via `_typos.toml` that excludes changelogs
and fixtures it's down to just a couple of typos, many of which are
false positives due to language, hashes or signatures which contain
two-letter sequences that it wants to correct. Especially with hashes
I find it hard to teach it what to do as I want the whole file to be
checked without allowing bogus words in an extended dictionary.