Cleaning up documentation
Applied this lint:

```
just 'clippy-flags=-W clippy::doc-markdown' clippy-fix
```

I also had to search/replace:

```
[`to_decimal`()]  ->  [`to_decimal()``]
```

I have never seen parenthesis used as part of the link, so this is perhaps why clippy handles it like that.

There are still a few warnings being produced, but they have to do with how some of the links are constructed, e.g. this link

```
[File::index_names().len()][File::index_names()]
```

shows a warning

> you should put bare URLs between `<`/`>` or make a proper Markdown link

Not certain how these should be fixed, but there are 10-20 of such warnings.  What are you linking to here?