fix release pipeline
The problem was that one of the `include` sections contained another 'match' field
which essentially removed os information from all other linux builds.

Further, linux musl builds don't work with openssl anymore, so we brutally disable
it for now, even though we could do better in the performance department if we wanted
to enabled faster zlib builds, but for now it was left for the future.