fix: no-want detection for negotiation phase is now consistent.
It being inconsistent was a reason for 'failing to parse server response' which
was empty as we didn't provide any wants to the server, but didn't detect that case
in the initial negotiation-preparation phase.

Turns out we didn't detect it as our special handling of implicit tags was not done
in the negotiation-preparation phase.

The fix consists of unifying the filtering phase to all places that needed, so
the preparation phase outcome is now consistent with what would have come later.