fix!: don't provide path to object-retrieval callback of `Pipeline::convert::to_git::IndexObjectFn()`.
It implies that one has to be ready to fetch any kind of path from the index, even though it's always the path to
the file that is currently converted.

Also fix a bug that could cause it to return input as unchanged even though it was read into a buffer already.