Differentiate between `Executable` and `Program`
It's somewhat confusing, but it seems sometimes `git` uses a shell, but other
times it doesn't. And that is differentiate by the type system.

It would still have to be (re)validated at some point if this is actually the case,
but I guess there was a reason for this in the first place.