fix!: rename `Repository::branch_remote_ref()` to `Repository::branch_remote_ref_name()`, add `direction` argument (also to `Repository::branch_remote_name()` and `Repository::branch_remote()`).
This better differentiates the return value from the corresponding ref objects,
which would require the named ref to exist in the repository.

The `direction` argument allows to get the reference to push to as well.
Further, it now takes a full ref name to support deriving the name of branches
to push to.

Regarding `Repository::branch_remote()`,  previously, this functionality
was only available from a `Reference`,
but now it's more generally available with just a branch name.

The method was also adjusted to permit looking up non-symbolic remote
names, like remotes that are specified by their URL.