remove "Copyright Sebastian Thiel ..." from LICENSE files
The "Copyright Sebastian Thiel ..." line in the LICENSE files confuse
our tooling and reviewers at Google, making it take longer than
necessary to import new versions of Gitoxide into our monorepo. IIUC,
the copyright is supposed to go in each source code file
instead. Maybe it will become necessary to move it out of the LICENSE
file as soon as the project decides to copy/vendor some code from
somewhere else anwyay, because then we probably can't claim copyright
for the entire repo? I'm not a lawyer and far from an export on
this.

If we remove the copyright line from the LICENSE files, does that mean
that we should add them in each source file? I think I've heard that
that's not necessary because the information about who wrote the code
is available in the commit log..