Fix stat on files with timestamps before the epoch
stat converted the seconds to a `u64` before converting them to
SystemTime. This would cause timestamps before the epoch to wrap around
to timestamps near `u64::MAX`.

Instead, pass the seconds as an i64. Convert it to SystemTime by
computing an absolute Duration and then adding or subtracting as
appropriate.