feat!: assure symlinks to directories are ignored with `dir/` declarations in `.gitignore`.
Initially, symlinks appear like symlinks thanks to `lstat`, but to do
exclude handling correctly these need another `stat` call.

However, this is also not done in Git, but done in `libgit2` only,
so a toggle was added to act like Git by default, but allow obtaining
the same behaviour as git2 for compatibility.