Add ambiguous path journey tests
This also adds a snapshot from running them, which shows the
expected failure. (However, in this commit the only snapshot that
is added is the one that was actually generated on the non-CI test
run. This pertains to the `clone` subcommand. Journey tests of the
`receive` subcommand are only run on CI, including the new test. It
should have a committed snapshot, but that's not included in this
commit.)

These differ from the ambiguous host tests, in that ambiguous hosts
are dangerous in the `gix` command when they appear in an `ssh://`
URL (with no `username@` preceding them), while ambiguous paths are
instead dangerous in the `gix` command when they appear in a
`username@host:path/repo` "URL".

The new tests should pass already, and the test of the `clone`
subcommand is confirmed to pass (as noted above, journey tests of
the `receive` subcommand are only run on CI). This is because
ambiguous hosts and ambiguous paths are already handled. In
contrast, ambiguous usernames, which have no journey tests yet, are
not yet handled in gix-transport.