Try, so far unsuccessfully, to add missing `-G` test
This is to test a code path where a host is passed to an ssh
command and it's important to ensure it doesn't start with `-`.
This code path uses `-G` to help figure out what SSH client is
being used, and most of the time it is not followed. The existing
check is sufficiently simple it is most likely correct. But it
seems never to to have been covered in automated tests: neither
unit tests nor any journey tests.

This does not yet manage to test is successfully, because it does
not succeed at getting that specific code path to be followed.