Add missing +x bit on scripts that are run and not sourced
This adds executable permissions to shell scripts that begin with a
`#!` and are meant to be run rather than sourced into another
script. Most of these already had those set, so the inconsistency
of having a handful of them unset does not appear intended.

The purpose of this change is to make clearer which .sh files are
meant to be run directly in a new shell process (which is most, but
not all, of them), and which are meant to be sourced into another
script rather than run (i.e., those that are script "libraries").
Executable permissions are intentionally not added to .sh files
without shebangs that are meant for sourcing rather than running.

This was not causing test failures because the way the scripts were
invoked did not rely on being able to directly execute them.
Nonetheless, it is clearest for these metadata to match the
scripts' shebangs. (Scripts that shouldn't have the executable bit
set should likewise not have shebangs, but there didn't appear to
be cases where scripts wrongly had shebangs or wrongly had +x set.)

A further benefit is that executable scripts are traditionally
named with no special suffix while files meant for sourcing into a
shell are traditionally named with a .sh suffix. Since that is not
the convention followed here, setting the executable bits may be
especially useful to clarify intent.

This commit also removes doubled end-of-file newlines (where a file
ended in two newline characters instead of just one) where present
in a few files whose permissions were being changed.

The rationale for including that here, even though no other script
cleanup is done here, is that sometimes this happens as an
overcorrection when editing scripts on Windows, where many editors
don't automatically add even a single end-of-file newline. Windows
is also where intended executable bits are sometimes omitted. That,
in turn, is due to the absence of a Unix-style chmod, and how
Cygwin-like environments infer likely executable permissions from
shebangs rather than file metadata. (git still tracks them, but
Windows's own permissions system differs).

Files whose metadata were not being changed were not checked for
extra newlines, nor otherwise edited, here.