Start on demo script making repo with .. trees, deploying above repo
This should not be incorporated into automated tests in its current
form. It is a proof of concept to generate repositories that attempt
to install real executables in directories where they may be run,
whereas test fixtures should completely limit all effects to testing
directories, even in the event of regressions or unexpected failures.