No need to actually create the directories
Because committing the staged paths creates the necessary Git tree
objects irrespective of what directories exist or are otherwise
represented.

In addition to simplifying the proof-of-concept repository, this
also makes it so its entries are properly ordered in its Git object
database, so `git fsck` does not report errors about that, and
exits reporting success (though of course still warns about the
presence of `..` components).