fix windows CI by trying installations of pure Rust version instead
Note that this doesn't truly capture the intent of the test anymore,
as the idea was to install the best-performing, default version.

But this broke due to ASM-hashes now failing on Windows and I reckon
that fixing it might be harder than one would think. So here is the quick-fix,
to get CI going again.