further testing of `.git` path variants
This is to see if anything should be done to more effectively
prevent paths containing `.git` (icase).

In conclusion, I think it's fine to keep allowing it as none
of the component-validations really kicks in on Linux if backslashes
are used as path separator. Thus, `.git` shouldn't be more special than
`..` for example.

The only way to fix this on Linux would be to either enable Windows protections,
or to disallow `\` as path seprator by default which seems too limitting.

Windows Users will naturally be protected as path-splitting will turn
these into components, with each of them checked as normal.