Regenerate archives for changed scripts
Changing the scripts' shebangs caused their CRC32 checksums, which
the test suite uses in deciding whether to use a pre-generated
archive instead of running them, to differ.

This was done by running `cargo nextest run --all --no-fail-fast`
on an Ubuntu 22.04 LTS system with Git 2.45.2 installed using the
package from https://launchpad.net/~git-core/+archive/ubuntu/ppa.
This resembles the ubuntu-latest CI image but is not identical to
it. All tests passed on that system, as expected. This commits the
modified (i.e. regenerated) archives.