fix: Check if folder is owned by administrators group (#1429)
The code had checked the process or thread's SID to see if it
matched the administrators group, but this will never be the case
since the process or thread is running as some user.

If the folder token is that of the administrators group, then at
that point we can check membership of the current thread SID in it.