Omit excess punctuation in `expect` messages
This removes the periods from the ends of sentences passed as
`expect` messages in recently introduced tests and their helpers,
because when called on a `Result`, `Some text: Result info` is
better than `Some text.: Result info`. Doing this also improves
stylistic consistency with other `expect` messages in the
project. I didn't make an analogous change to assertion messages,
because the awkward `.:` construction does not arise from them.

There may be further refinement that can be done, as the new
`expect` messages are still worded as assertions and, as such,
could be misread as claiming that the situation they decribe
wrongly happened, rather than the intended meaning that it wrongly
did not happen.

This also changes "key" to "registry key" in one of the `expect`
messages so that its meaning is clear in output of big test runs.