Use bash for everything on all platforms in release workflow
The way environment variables are being set and accessed assumes
a Bourne-style (POSIX-like) shell, but PowerShell was being used on
Windows for some steps where this was being assumed.

It's possible to set bash only for some steps, but that is more
complex to understand than using it for everything.