Write universal binary build job draft; refactor env vars
This expands the stub job for making Universal 2 binaries for macOS
into a full definition, downloading and extracting the two
architecture-specific archives from the GitHub release. This may
require further refinement.

It also refactors the way environment variables are set for the
preexisting jobs:

- Order variables in `env:` so they may be easier to understand.
- Define `VERSION` in `env:` rather than in its own step.

I think this is a bit more readable, but also, it allows the new
job to be stylistically consistent with the preexisting jobs.