Unset GIT_ASKPASS in tests that assume it is unset
Two of the askpass tests assume `GIT_ASKPASS` is not set. Before
this change, they wrongly fail when it is set, even if the code
under test is working correctly.

1. ssh_askpass_is_used_as_fallback
2. ssh_askpass_does_not_override_current_value

As other tests that deliberately involve `GIT_ASKPASS` (and
therefore do not need to be modified) attest, when it is set:

1. `SSH_ASKPASS` is not used as a fallback (`GIT_ASKPASS`
   deliberately takes precedence).

2. Although `SSH_ASKPASS` still does not override the current
   value, `GIT_ASKPASS` does override it, so checking for equality
   to the current value fails.

Since those test cases are among those that already have the
`serial` attribute and already temporarily modify the environment,
this extends that modification by having it also temporarily unset
`GIT_ASKPASS` if it was set.