Run CI tests for push event on more branches
When a branch is named with a non-trailing (i.e. directory-like)
`run-ci` component, such as `run-ci/foo` or `bar/run-ci/baz`,
pushing to the branch will run the CI test workflow under the same
conditions that a push to `main` would run it.

This is primarily to allow CI to work in forks even when there is
no pull request, while still avoiding running it when it is not
wanted. Note that forks still need to enable GitHub Actions
workflows for this, or any workflows, to run in them.