Drop `once_cell::sync::Lazy` in favor of `std::sync::LazyLock`
This almost completely drops the `once_cell` crate, but `OnceCell` is
still in use. While the `OnceCell` type is available in the standard
library, the `get_or_try_init` method is not yet stabilized.

This bumps the MSRV up to 1.80, which is the first version to stabilize
`std::sync::LazyLock`.