Workspace Clippy lint management
Make all lints managed centrally in one place.

Note that this approach forces pedantic lints on the entire codebase - which means that when the new versions of rust are released, some new lint may fail the build.  At the same time, this allows new lints to always be noticed and manually excluded.

The big list of allowed lints can be slowly worked through and forbidden one by one.