Fix EXE_INFO no local scope test for macOS
By setting `GIT_CONFIG_NOSYSTEM` (to `1`), rather than setting
`GIT_CONFIG_SYSTEM` (to `/dev/null`). The test still sets
`GIT_CONFIG_GLOBAL` to `/dev/null.

Although setting `GIT_CONFIG_SYSTEM` to `/dev/null` (or even to a
file with actual configuration in it) does not take precedence over
the "unknown" scope (usually) under `/Library`, and documentation
in git(1) and git-config(1) seems to suggest `GIT_CONFIG_SYSTEM`
and `GIT_CONFIG_NOSYSTEM` would affect the same scope(s), it turns
out that setting `GIT_CONFIG_NOSYSTEM` to `1` is sufficient to
cause both the system scope and the installation-associated
"unknown" scope to be omitted from `git config -l` output.

This fixes both the false positive and false negative in that test
case, which is now as precise on macOS as on other systems. In
addition, because `GIT_CONFIG_NOSYSTEM` can be set in real-world
scenarios (to make a configuration reproducible or to suppress
broken or wanted systemwide configuration, as documented), the bug
where a local configuration file path could be used affected macOS
more fully than I had anticipated.

This commit also improves a variable name and adds an assertion
message to clarify the test.