Check that the test affects `env::temp_dir()` as desired
This check is probably overzealous in that a path that identifies
the same directory in a different way that is non-equivalent under
path equality would also be acceptable. But a more restrictive
check is simpler, and since we have canonicalized the path and used
it after that for both changing the directory and setting the
environment variables we intend that `env::temp_dir()` will use,
that is unlikely to be a problem.

That it not to say that this cannot break in practice. Rather, it
can break, but if it does, there is a substantial likelihood that
the test is not ensuring the behavior it wishes to check. So to
preserve it as a regression test, failures of this new assertion
should be examined.

This commit also removes some old cruft (commented out test code I
had used while investigating a test bug) and rewords some custom
assertion messages so it is clearer what the expectation is.