Add tests for `gix_testtools::Env`
The `nonoverlapping` test, which fortunately is what most closely
resembles existing known usage (and most likely all current usage
in gitoxide's own test suite), already passes.

The other tests test the situation where the same environment
variable is affected by multiple `set` or `unset` calls (or both a
`set` and an `unset` call). These do not pass yet, because while
the assertions about the immediate effect on the environment of
each such call all pass, the assertions about the effect after
drop fail.

This is because, on drop, `Env` currently restores the state of a
variable that was most recently saved, i.e., it puts it back to
whatever it was just before the most recent modification it made to
it.

This goes against the intuitive expectation that `Env` will reset
things to the way they were before the `Env` object was created and
used (so long as all changes were by `set` and `unset` calls).