fix: Let `gix_testtools::Env` undo multiple changes to the same var
Previously, an `Env` instance would restore the original state on
drop if no more than one modification was made to any one variable
through it, but would restore an intermediate state if the same
variable was ever set multiple times, unset multiple times, or both
set and unset in any order.

The state it would restore for each variable was its state
immediately before the most recent modification (through the `Env`
instance) that affected it, rather than its original state before
the first time it was modified through that `Env` instance.

This fixes that by undoing the changes in the opposite of the order
they were made.