Don't assert that `tempfile::TempDir` cleans up
It still does clean up (that is done by RAII; explicitly calling
`close()` is just to be able to identify and handle errors). The
assertion that it does is not part of what the test is testing, nor
is it critical to the correctness of the test itself. I had put it
there to catch bugs that might arise during development of the test
itself. (In some libraries/frameworks, a recursive deletion on
Windows that involves files that need to be accessed with `\\?\`
paths does not automatically succeed.)