Rename to `GIT_HIGHEST_SCOPE_CONFIG_PATH`
This renames the crate-internal static item with the cached path
of the highest scope nonempty config, treated as associated with
the `git` installation, from `GIT_HIGHEST_PRIORITY_CONFIG_PATH` to
`GIT_HIGHEST_SCOPE_CONFIG_PATH`.

This reflects the propensity of "priority" to have the opposite of
the intended meaning here, since configuration in this file stands
to be superseded by configuration in any lower (i.e. any other)
scope. See #1568 for context.

I think this change from "PRIORITY" to "SCOPE" in the name
preserves the benefits of the change from `EXE_INFO` to
`GIT_HIGHEST_PRIORITY_CONFIG_PATH` (dd2d666), while making clearer
what is going on and avoiding misinterpretations.