Add missing executable bits
Some of the more recently introduced fixture scripts were not
marked executable, even though the others are.

Because `gix-testtools` has a fallback that runs scripts with
`bash` explictly when executing them directly fails, this did not
break any scripts, but the executable bit is also useful to
document which files are meant to be runnable scripts.

This marks them executable. Now all the shell scripts that begin
with a `#!` have the executable bit set. (The handful of `.sh`
files that are not meant to be run as scripts, but instead are
sourced by other scripts, do not start with `#!` and continue not
to have `+x` applied to them.)