Rename create_archive_if_{not_on_ci -> we_should}
This renames the private funciton `create_archive_if_not_on_ci` in
gix-testtols to `create_archive_if_we_should`, because it already
checks for much more than whether we are on CI to decide whether
or not to create the archive. This archive is also (already) not
created if we are running on Windows, if the archive destination is
excluded by being ignored in any `.gitignore`, or if Git LFS is in
use and the archive is an LFS pointer file.