fix: keep multi-line header value newlines verbatim
Previously they were normalized to \n which made round-trips
degenerate information.

It's notable that we explicitly do NOT normalize the newline
at the very last line of the multi-line header as its part
of that line, even though it's also relevant to make the header field
valid. The reason is that these *can* also be written with \r\n, and
we don't want to degenerate that information.