fix: properly classify non-bare repositories without worktree.
Such repositories claim they are not bare, but also don't have
a workdir with a .git directory inside of it (which is always needed
unless certain variables are set).

Now they are classified as repository that needs more detailed checking
later, while indicating that it 'usually' would be having a worktree
that is configured via git configuration.