doc: Fix some broken links to commit/blob URLs
A few of the links discovered to have been broken while updating
them are easy to fix. This fixes those.

They were broken due to changes in the repository structure or
contents that were not fully reflected when the links were last
updated. For example, when `git-*` crates were renamed to their
current `gix-*` names, URLs were updated, but some of these were
GitHub URLs where the commit hash was for a commit prior to the
rename.